Phase 3f: Extract router logic to serv/routing/router.py - PHASE 3 COMPLETE
- Extracted complete Router class (~360 lines) from serv/_routing.py
- Moved RouteSettings class and get_current_router function
- Includes route registration, mounting, URL generation, resolution logic
- Updated serv/_routing.py to re-export module for backward compatibility
- Enhanced routing __init__.py to export all routing utilities
- All 298 tests passing, no functionality changes
- Maintains full backward compatibility through re-exports

🏁 PHASE 3 COMPLETE: Full routing layer modularization achieved
✅ Patterns, decorators, generation, resolution, handlers, router all extracted
✅ Clean separation of concerns with focused modules
✅ Zero breaking changes, full backward compatibility
